core: probe: kernel: move the inspection logic to its own file
The only functional change is the BTF file used for tests is now loaded
using #[cfg(test)] instead of manually.
